Move operation node codes to separate file
