+++ nodedge/blocks/autogen/operator/mul_block.py
@@ -12,7 +12,7 @@ from nodedge.connector import SocketType
 _LOG = logging.getLogger(__name__)
 
 try:
-    from nodedge.blocks.block_config import OP_NODE_MULTIPLY
+    from nodedge.blocks.op_node import OP_NODE_MULTIPLY
 except NameError:
     _LOG.warning(f"Not registered block: {__name__}")
     op_block_string = -1
