+++ nodedge/blocks/custom/output_block.py
@@ -2,7 +2,8 @@
 from typing import List
 
 from nodedge.blocks.block import Block
-from nodedge.blocks.block_config import BLOCKS_ICONS_PATH, OP_NODE_OUTPUT, registerNode
+from nodedge.blocks.block_config import BLOCKS_ICONS_PATH, registerNode
+from nodedge.blocks.op_node import OP_NODE_OUTPUT
 from nodedge.blocks.block_exception import EvaluationError
 from nodedge.blocks.graphics_block import GraphicsBlock
 from nodedge.blocks.graphics_output_block_content import GraphicsOutputBlockContent
