+++ nodedge/blocks/op_node.py
@@ -1,0 +1,15 @@
+OP_NODE_INPUT = 1
+OP_NODE_OUTPUT = 2
+OP_NODE_ADD = 3
+OP_NODE_SUBTRACT = 4
+OP_NODE_MULTIPLY = 5
+OP_NODE_DIVIDE = 6
+OP_NODE_MODULO = 7
+OP_NODE_POWER = 8
+OP_NODE_LESS = 9
+OP_NODE_LESS_EQUAL = 10
+OP_NODE_EQUAL = 11
+OP_NODE_GREATER = 12
+OP_NODE_GREATER_EQUAL = 13
+OP_NODE_FLOOR_DIVIDE = 14
+OP_NODE_COMPUTE_ABSOLUTE = 15
