+++ tools/block_generator/block_template.txt
@@ -12,7 +12,7 @@ from nodedge.connector import SocketType
 _LOG = logging.getLogger(__name__)
 
 try:
-    from nodedge.blocks.block_config import ${op_block_string}
+    from nodedge.blocks.op_node import ${op_block_string}
 except NameError:
     _LOG.warning(f"Not registered block: {__name__}")
     op_block_string = -1
